Flyttet gjenbrukskode ut i en egen GjenbrukVilkårService
