+++ src/main/kotlin/no/nav/familie/ef/sak/vilkår/VurderingController.kt
@@ -7,6 +7,7 @@ import no.nav.familie.ef.sak.vilkår.dto.OppdaterVilkårsvurderingDto
 import no.nav.familie.ef.sak.vilkår.dto.SvarPåVurderingerDto
 import no.nav.familie.ef.sak.vilkår.dto.VilkårDto
 import no.nav.familie.ef.sak.vilkår.dto.VilkårsvurderingDto
+import no.nav.familie.ef.sak.vilkår.gjenbruk.GjenbrukVilkårService
 import no.nav.familie.ef.sak.vilkår.regler.Vilkårsregler
 import no.nav.familie.kontrakter.felles.Ressurs
 import no.nav.familie.kontrakter.felles.objectMapper
@@ -29,7 +30,8 @@ import java.util.UUID
 class VurderingController(
     private val vurderingService: VurderingService,
     private val vurderingStegService: VurderingStegService,
-    private val tilgangService: TilgangService
+    private val tilgangService: TilgangService,
+    private val gjenbrukVilkårService: GjenbrukVilkårService
 ) {
 
     private val secureLogger = LoggerFactory.getLogger("secureLogger")
@@ -88,7 +90,7 @@ class VurderingController(
         tilgangService.validerTilgangTilBehandling(request.kopierBehandlingId, AuditLoggerEvent.ACCESS)
         tilgangService.validerTilgangTilBehandling(request.behandlingId, AuditLoggerEvent.UPDATE)
         tilgangService.validerHarSaksbehandlerrolle()
-        vurderingService.gjenbrukInngangsvilkårVurderinger(request.behandlingId, request.kopierBehandlingId)
+        gjenbrukVilkårService.gjenbrukInngangsvilkårVurderinger(request.behandlingId, request.kopierBehandlingId)
         return Ressurs.success(vurderingService.hentEllerOpprettVurderinger(request.behandlingId))
     }
 }
